Create menu pipeline

+++ CleanDDDinner.Domain/Common/ValueObjects/AverageRating.cs
@@ -4,13 +4,13 @@ namespace CleanDDDinner.Domain.Common.ValueObjects;
 
 public sealed class AverageRating : ValueObject
 {
-    public double Value { get; private set; }
+    public float Value { get; private set; }
     public int NumRatings { get; private set; }
 
-    private AverageRating(double value, int numRatings)
+    private AverageRating(float value, int numRatings)
         => (Value, NumRatings) = (value, numRatings);
 
-    public static AverageRating Create(double rating = 0, int numRatings = 0)
+    public static AverageRating Create(float rating = 0, int numRatings = 0)
         => new(rating, numRatings);
 
     public void AddNewRating(Rating rating)
